test/requirements.txt
azure.mgmt.resource
requests
pytest
pytest-cov
paramiko
scp